+++ src/mission_control/src/store.rs
@@ -1,5 +1,5 @@
 use crate::STATE;
-use shared::types::interface::UserId;
+use shared::types::state::UserId;
 
 pub fn get_user() -> UserId {
     STATE.with(|state| state.borrow().stable.user.user).unwrap()
